Added additional layer to end of MAE + minor changes to sampling
In order to increase model flexibility an additional FC layer was added
at the end of the MAE model (e.g. right before outputing the predicted value
embedding). In the future, the number of output layers should be made an
additional parameter in the mae_config.yaml file.

A couple changes were made to the sampling code.

1. No longer going through all possible attributes when generating evaluation
    samples. This was needed to quickly generate output examples for the SoCal
    NLP talk (which did not discuss the work done on <UNK> predictions). Future
    versions should have the flexibility to toggle which kind of evaluation to
    perform.
2. When generating batches a list of product URIs is now emitted in addition to
    the feed_dict. This makes it much easier to match model outputs to the
    corresponding products in the MAE dataset.
3. Corrupted images were not being omitted when generating samples for
    evaluation. Excpetion handling was added to address this issue.